--- packages/repack/src/plugins/ChunksToHermesBytecodePlugin/index.ts
@@ -1,1 +1,0 @@
-export * from './ChunksToHermesBytecodePlugin.js';
