+++ packages/repack/src/plugins/HermesBytecodePlugin/index.ts
@@ -1,0 +1,1 @@
+export * from './HermesBytecodePlugin.js';
