+++ packages/repack/src/plugins/HermesBytecodePlugin/utils/index.ts
