+++ packages/repack/src/plugins/HermesBytecodePlugin/utils/transformBundleToHermesBytecode.ts
